promoted app and account generators to utils
