+++ config/webtesting_prod.conf
@@ -12,6 +12,8 @@ plan_type=Basic
 username=nsolaiappan@crittercism.com
 password=CritPass123
 test_user_password=CritPass123
+basic_username = "tkerbosch+integration@crittercism.com"
+basic_password = "crittercism"
 
 test_user_engg    = crittercism.engg@yahoo.com
 test_user_manager = critterrism.manager@gmail.com
