+++ config/webtesting_regression_test-stage.conf
@@ -11,6 +11,9 @@ plan_type = Enterprise
 [login]
 username=nsolaiappan@crittercism.com
 password=CritPass123
+basic_username = "tkerbosch+integration@crittercism.com"
+basic_password = "crittercism"
+
 test_user_password= CritPass123
 test_user_engg    = crittercism.engg@yahoo.com
 test_user_manager = critterrism.manager@gmail.com
